pythonProject3/requirements.txt
Flask==2.2.2
# pillow-heif==0.9.0
Pillow==9.3.0
opencv-python==4.6.0.66
numpy==1.21.0
Clustimage
scikit-learn==1.2.0
torchvision==0.8.2
matplotlib==3.3.3
torch==1.7.1
minio==7.1.13